+++ src/controllers/user.py
@@ -9,7 +9,7 @@ from ..libs.constraints import FIELD_PAGE_NUMBER, FIELD_PAGE_SIZE, FIELD_STRING_
 from ..libs.enum import AuthorityEnum
 from ..libs.openapi_tags import TagNameEnum
 from ..libs.page import Page
-from ..services.auth import UserDepends
+from ..services.authorize import UserDepends
 from ..usecases.user import UserUsecase
 
 router = APIRouter()
