+++ src/dto/auth.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel, ConfigDict
 
 from ..libs.constraints import FIELD_STRING_USERNAME
 from ..libs.enum import AuthorityEnum
-from ..services.auth import Token
+from ..services.authorize import Token
 
 
 ## ログインレスポンス
