+++ src/services/base.py
@@ -1,6 +1,5 @@
-from sqlalchemy.orm.session import Session
-
 from ..libs.config import get_config
+from ..entities.user import UserEntity
 
 
 class ServiceError(Exception):
@@ -12,21 +11,4 @@ class ServiceBase:
     サービス基底クラス
     """
 
-    class Error(ServiceError):
-        """
-        サービスエラー
-        """
-
-        pass
-
-    def __init__(self, session: Session) -> None:
-        """
-        サービスクラスの初期化
-
-        Args:
-            session: データベースセッション
-        """
-        self.session = session
-        "セッション"
-        self.config = get_config()
-        "設定値"
+    config = get_config()
